+++ packages/htmlgoddess-cli/src/commands/serve/index.ts
@@ -1,5 +1,4 @@
 import { Command, flags } from "@oclif/command";
-import HTMLGoddess from "htmlgoddess";
 import execa from "execa";
 import { CWD_PATH } from "../../index";
 import path from "path";