+++ CHANGELOG.md
@@ -1,3 +1,7 @@
 
 0.0.1
 - Abstract core htmlgoddess bin from CLI commands.
+
+0.0.2 
+- fix: git save/publish commands not cleaning up.
+- fix: global cli not loading dependancies.